in PE/COFF, support relocations relative to symbol
in ELF, there is a relocation type where the calculated address equals
the address of a given symbol, plus a given addend relative to the
symbol's address. I haven't found a way to achieve that in PE/COFF. So
we now emulate it by creating unnamed symbols and calculating the offset
manually.

Unfortunately, this only works for defined symbols whose address is
known when the .obj file is written. That is sufficient for now since
our only use of function-internal jumps. (These should be replaced later
by relative jumps anyway, to not need a relocation at all).

What should we do for symbol+addend relocations for undefined symbols? I
imagine we will need those later, to access global data.