implement value macros
Value macros are like normal macros (which are now called function
macros) with no parameters, with the difference that value macros are
not called with parentheses but just as a value reference (EXPR_SYMREF).
Macros are now very much like C's #define (which also has a value and a
function variant), but at the syntactic level instead of lexical.

It's not clear to me if value macros are a good idea, since they can
still access mutable data which may come as a surprise to the user. But
we'll try. Probably it's good idea to trust the programmer to do the
right thing, and in this case value macros also remove the need for a
"const" facility.