try workaround to get BSS to work under PECOFF
PECOFF is just unlogical. We're doing a swapping of fields now that
seems completely unnecessary. But it seems that for BSS in PECOFF,
"VirtualSize" (size in running program) really means "size on disk", and
the other way around. Also, PECOFF requires the Pointer-to-Data to be 0
even though there is no data (so it seems like a needless requirement
even though shouldn't matter)