improve inference of constants and types.
Inference of constants and types is a little complicated. Constants and
types are interdependent since array lengths are constants. This commit
tries to improve the inference code. At least one problem still remains:
We traverse the AST again (which means coupling), while the only
important thing is to infer constants and types in the correct order.